prompt files: only show problems for files open in editor (#304872)
* prompt files: only show problems for files open in editor

* update
