+++ src/vs/workbench/contrib/chat/common/promptSyntax/languageProviders/promptValidator.ts
@@ -6,21 +6,15 @@
 import { isEmptyPattern, parse, splitGlobAware } from '../../../../../../base/common/glob.js';
 import { Iterable } from '../../../../../../base/common/iterator.js';
 import { Range } from '../../../../../../editor/common/core/range.js';
-import { ITextModel } from '../../../../../../editor/common/model.js';
-import { IModelService } from '../../../../../../editor/common/services/model.js';
 import { localize } from '../../../../../../nls.js';
 import { IConfigurationService } from '../../../../../../platform/configuration/common/configuration.js';
-import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
-import { IMarkerData, IMarkerService, MarkerSeverity } from '../../../../../../platform/markers/common/markers.js';
+import { IMarkerData, MarkerSeverity } from '../../../../../../platform/markers/common/markers.js';
 import { ChatMode, IChatMode, IChatModeService } from '../../chatModes.js';
 import { ChatConfiguration, ChatModeKind } from '../../constants.js';
 import { ILanguageModelChatMetadata, ILanguageModelsService } from '../../languageModels.js';
 import { ILanguageModelToolsService, SpecedToolAliases } from '../../tools/languageModelToolsService.js';
-import { getPromptsTypeForLanguageId, PromptsType, Target } from '../promptTypes.js';
+import { PromptsType, Target } from '../promptTypes.js';
 import { ISequenceValue, IHeaderAttribute, IScalarValue, parseCommaSeparatedList, ParsedPromptFile, PromptHeader, IValue, PromptHeaderAttributes } from '../promptFileParser.js';
-import { Disposable, DisposableStore, toDisposable } from '../../../../../../base/common/lifecycle.js';
-import { Delayer } from '../../../../../../base/common/async.js';
-import { ResourceMap } from '../../../../../../base/common/map.js';
 import { IFileService } from '../../../../../../platform/files/common/files.js';
 import { IPromptsService } from '../service/promptsService.js';
 import { ILabelService } from '../../../../../../platform/label/common/label.js';
@@ -1229,103 +1223,4 @@ export function getTarget(promptType: PromptsType, header: PromptHeader | URI): 
 
 function toMarker(message: string, range: Range, severity = MarkerSeverity.Error): IMarkerData {
 	return { severity, message, ...range };
-}
-
-export class PromptValidatorContribution extends Disposable {
-
-	private readonly validator: PromptValidator;
-	private readonly localDisposables = this._register(new DisposableStore());
-
-	constructor(
-		@IModelService private modelService: IModelService,
-		@IInstantiationService instantiationService: IInstantiationService,
-		@IMarkerService private readonly markerService: IMarkerService,
-		@IPromptsService private readonly promptsService: IPromptsService,
-		@ILanguageModelsService private readonly languageModelsService: ILanguageModelsService,
-		@ILanguageModelToolsService private readonly languageModelToolsService: ILanguageModelToolsService,
-		@IChatModeService private readonly chatModeService: IChatModeService,
-	) {
-		super();
-		this.validator = instantiationService.createInstance(PromptValidator);
-
-		this.updateRegistration();
-	}
-
-	updateRegistration(): void {
-		this.localDisposables.clear();
-		const trackers = new ResourceMap<ModelTracker>();
-		this.localDisposables.add(toDisposable(() => {
-			trackers.forEach(tracker => tracker.dispose());
-			trackers.clear();
-		}));
-		this.modelService.getModels().forEach(model => {
-			const promptType = getPromptsTypeForLanguageId(model.getLanguageId());
-			if (promptType) {
-				trackers.set(model.uri, new ModelTracker(model, promptType, this.validator, this.promptsService, this.markerService));
-			}
-		});
-
-		this.localDisposables.add(this.modelService.onModelAdded((model) => {
-			const promptType = getPromptsTypeForLanguageId(model.getLanguageId());
-			if (promptType && !trackers.has(model.uri)) {
-				trackers.set(model.uri, new ModelTracker(model, promptType, this.validator, this.promptsService, this.markerService));
-			}
-		}));
-		this.localDisposables.add(this.modelService.onModelRemoved((model) => {
-			const tracker = trackers.get(model.uri);
-			if (tracker) {
-				tracker.dispose();
-				trackers.delete(model.uri);
-			}
-		}));
-		this.localDisposables.add(this.modelService.onModelLanguageChanged((event) => {
-			const { model } = event;
-			const tracker = trackers.get(model.uri);
-			if (tracker) {
-				tracker.dispose();
-				trackers.delete(model.uri);
-			}
-			const promptType = getPromptsTypeForLanguageId(model.getLanguageId());
-			if (promptType) {
-				trackers.set(model.uri, new ModelTracker(model, promptType, this.validator, this.promptsService, this.markerService));
-			}
-		}));
-
-		const validateAll = (): void => trackers.forEach(tracker => tracker.validate());
-		this.localDisposables.add(this.languageModelToolsService.onDidChangeTools(() => validateAll()));
-		this.localDisposables.add(this.chatModeService.onDidChangeChatModes(() => validateAll()));
-		this.localDisposables.add(this.languageModelsService.onDidChangeLanguageModels(() => validateAll()));
-	}
-}
-
-class ModelTracker extends Disposable {
-
-	private readonly delayer: Delayer<void>;
-
-	constructor(
-		private readonly textModel: ITextModel,
-		private readonly promptType: PromptsType,
-		private readonly validator: PromptValidator,
-		@IPromptsService private readonly promptsService: IPromptsService,
-		@IMarkerService private readonly markerService: IMarkerService,
-	) {
-		super();
-		this.delayer = this._register(new Delayer<void>(200));
-		this._register(textModel.onDidChangeContent(() => this.validate()));
-		this.validate();
-	}
-
-	public validate(): void {
-		this.delayer.trigger(async () => {
-			const markers: IMarkerData[] = [];
-			const ast = this.promptsService.getParsedPromptFile(this.textModel);
-			await this.validator.validate(ast, this.promptType, m => markers.push(m));
-			this.markerService.changeOne(MARKERS_OWNER_ID, this.textModel.uri, markers);
-		});
-	}
-
-	public override dispose() {
-		this.markerService.remove(MARKERS_OWNER_ID, [this.textModel.uri]);
-		super.dispose();
-	}
 }
